Related work items: #14, #15, #16, #17, #20, #22, #23, #25, #26, #130, #179, #180, #181, #182, #183, #184, #185, #186, #187, #188, #189, #191, #192, #193, #194, #195, #196, #197, #200, #220, #221, #230, #247, #254, #255, #264, #265

+++ WhatIfBudget/WhatIfBudget.Logic/BudgetLogic.cs
@@ -24,6 +24,9 @@ namespace WhatIfBudget.Logic
         private readonly ISavingGoalService _savingGoalService;
         private readonly IDebtGoalService _debtGoalService;
         private readonly IMortgageGoalService _mortgageGoalService;
+
+        private readonly IIncomeLogic _incomeLogic;
+        private readonly IExpenseLogic _expenseLogic;
         
         public BudgetLogic(IBudgetService budgetService,
                            IIncomeService incomeService,
@@ -35,7 +38,9 @@ namespace WhatIfBudget.Logic
                            IInvestmentGoalInvestmentService investmentGoalInvestmentService,
                            IMortgageGoalService mortgageGoalService,
                            IDebtGoalService debtGoalService,
-                           ISavingGoalService savingGoalService) { 
+                           ISavingGoalService savingGoalService,
+                           IIncomeLogic incomeLogic,
+                           IExpenseLogic expenseLogic) { 
             _budgetService = budgetService;
             _incomeService = incomeService;
             _expenseService = expenseService;
@@ -47,6 +52,8 @@ namespace WhatIfBudget.Logic
             _savingGoalService = savingGoalService;
             _debtGoalService = debtGoalService;
             _mortgageGoalService = mortgageGoalService;
+            _incomeLogic = incomeLogic;
+            _expenseLogic = expenseLogic;
         }
 
         public IList<UserBudget> GetUserBudgets(Guid userId)
@@ -67,15 +74,31 @@ namespace WhatIfBudget.Logic
             else { return dbBudget; }
         }
 
-        public double GetSumOfGoalAllocations(int budgetId)
+        public double GetAvailableMonthlyNet(int budgetId)
         {
-            // TODO: complete this as other goal services are fleshed out.
+            // Get net of user-entered incomes and expenses
+            var monthlyNet = _incomeLogic.GetBudgetMonthlyIncome(budgetId);
+            monthlyNet -= _expenseLogic.GetBudgetMonthlyExpense(budgetId);
 
-            var dbInvestmentGoal = _investmentGoalService.GetInvestmentGoal(budgetId);
+            var dbBudget = _budgetService.GetAllBudgets().Where(x => x.Id == budgetId).FirstOrDefault();
+            if (dbBudget == null) { throw new NullReferenceException(); }
+
+            // Subtract existing goal allocations
+            var dbSavingGoal = _savingGoalService.GetSavingGoal(dbBudget.SavingGoalId);
+            if (dbSavingGoal == null) { throw new NullReferenceException(); }
+            monthlyNet -= dbSavingGoal.AdditionalBudgetAllocation;
+
+            var dbDebtGoal = _debtGoalService.GetDebtGoal(dbBudget.DebtGoalId);
+            monthlyNet -= dbDebtGoal.AdditionalBudgetAllocation;
+
+            var dbMortgageGoal = _mortgageGoalService.GetMortgageGoal(dbBudget.MortgageGoalId);
+            monthlyNet -= dbMortgageGoal.AdditionalBudgetAllocation;
+
+            var dbInvestmentGoal = _investmentGoalService.GetInvestmentGoal(dbBudget.InvestmentGoalId);
             if (dbInvestmentGoal == null) { throw new NullReferenceException(); }
-            UserInvestmentGoal investmentGoal = UserInvestmentGoal.FromInvestmentGoal(dbInvestmentGoal);
+            monthlyNet -= dbInvestmentGoal.AdditionalBudgetAllocation;
 
-            return investmentGoal.additionalBudgetAllocation /* + savingGoal.additionalBudgetAllocation...*/;
+            return monthlyNet;
         }
 
         public UserBudget? CreateUserBudget(Guid userId, UserBudget budget)
@@ -152,10 +175,12 @@ namespace WhatIfBudget.Logic
             // Delete associated goals
             var dbIG = _investmentGoalService.DeleteInvestmentGoal(budget.InvestmentGoalId);
             if (dbIG == null) { throw new NullReferenceException(); }
+
+            var dbSG = _savingGoalService.DeleteSavingGoal(budget.SavingGoalId);
+            if (dbSG == null) { throw new NullReferenceException(); }
             /*
-            dbSG = _savingGoalService.DeleteSavingGoal(budget.SavingGoalId);
-            dbDG = _debtGoalService.DeleteDebtGoal(budget.DebtGoalId);
-            dbMG = _mortgageGoalService.DeleteMortageGoal(budget.MortageGoalId);
+            var dbDG = _debtGoalService.DeleteDebtGoal(budget.DebtGoalId);
+            var dbMG = _mortgageGoalService.DeleteMortageGoal(budget.MortageGoalId);
             */
 
             // Delete associated incomes
